refactor: rename entity, update token instanse, throw error
