+++ .gitignore
@@ -53,3 +53,4 @@ Thumbs.db
 scripts/errors_*
 scripts/archive
 coverage
+project.yaml
