+++ docker-compose.yml
@@ -25,9 +25,9 @@ services:
       NODE_ENV: local
       DB_HOST: postgres
       DB_PORT: 5432
-      START_BLOCK: 1362830
-      NETWORK_ENDPOINT: wss://testnet-rpc.polymesh.live
-      # NETWORK_ENDPOINT: ws://host.docker.internal:9944
+      # START_BLOCK: 1362830
+      # NETWORK_ENDPOINT: wss://testnet-rpc.polymesh.live
+      NETWORK_ENDPOINT: ws://host.docker.internal:9944
     command:
       - --batch-size=500
       - -f=/app
