+++ schema.graphql
@@ -66,17 +66,17 @@ type Extrinsic @entity {
   spec_version_id: Int!
 }
 
-type TokenHolder @jsonField {
+type AssetHolder @jsonField {
   did: String!
   amount: String!
 }
 
-type TokenDocument @jsonField {
+type AssetDocument @jsonField {
   name: String!
   link: String!
 }
 
-type Token @entity {
+type Asset @entity {
   id: ID! # ticker
   ticker: String! @index(unique: true)
   name: String
@@ -85,11 +85,11 @@ type Token @entity {
   isDivisible: Boolean!
   isFrozen: Boolean!
   isUniquenessRequired: Boolean!
-  documents: [TokenDocument]!
+  documents: [AssetDocument]!
   identifiers: String! # JsonString
   ownerDid: String!
   fullAgents: [String!]!
-  holders: [TokenHolder]!
+  holders: [AssetHolder]!
   totalSupply: String!
   # totalDistributions: String!
   # totalTransfers: String!
